pss, The PSS Analysis Algorithm has been entirely reviewed
  The PSS Analysis Algorithm has been entirely reviewed and now
   it works very well. Some additional files have been modified,
   like configure.ac, plotting strategies and so on...

+++ src/spicelib/analysis/pssaskq.c
@@ -35,7 +35,7 @@ PSSaskQuest(CKTcircuit *ckt, JOB *anal, int which, IFvalue *value)
         }
         break;
     case PSS_POINTS:
-        value->iValue = job->PSSpoints;
+        value->iValue = (int)job->PSSpoints;
         break;
     case PSS_HARMS:
         value->iValue = job->PSSharms;
